Changes to try to fix merge
