+++ .idea/.gitignore
@@ -2,3 +2,4 @@
 # Default ignored files
 /shelf/
 /workspace.xml
+/.idea/
