+++ src/application/views/math_screen.py
@@ -26,7 +26,7 @@ class Questions:
         self.ID = ID
         self.all_questions_taken = []
         # self.all_questions_taken.append([f"What is {self.first_number} + {self.second_number}?",  f"{self.answer}"])
-        # print(self.all_questions_taken)
+        print(self.all_questions_taken)
 
     def toggle_topics(self):
         # Addition
@@ -112,13 +112,13 @@ class Questions:
             self.answer = int(self.first_number / self.second_number)
             return f"What is {self.first_number} / {self.second_number}?"
 
+
 # TODO: Prevent same question from appearing multiple times
-
-
 class Math_Screen(tk.Frame):
     def __init__(self, parent, ID, *args, **kwargs):
         super().__init__(parent, *args, **kwargs)
 
+        self.parent = parent
         self.answer_verification = tk.StringVar()
         self.ans_insert = tk.StringVar()
         self.insert_num = tk.StringVar()
@@ -126,6 +126,8 @@ class Math_Screen(tk.Frame):
         self.Total_Questions = 3
         self.questions_list = []
         self.finished = False
+
+        self.Correct_Answers = 0
 
         self.Question_label = tk.StringVar()
         self.Question_label.set(f"Question # of {self.Total_Questions}")
@@ -172,6 +174,8 @@ class Math_Screen(tk.Frame):
         self.number_button9 = ttk.Button(self, text="9", command=lambda: self.UserInsert_entry.insert('end', "9"))
         self.decimal_button = ttk.Button(self, text=".", command=lambda: self.UserInsert_entry.insert('end', "."))
 
+
+
         # Grid Layout
         self.UserInsert_entry.grid(row=5, column=2, sticky=(tk.E))
         self.submit_button.grid(row=5, column=3, sticky=tk.E)
@@ -193,9 +197,13 @@ class Math_Screen(tk.Frame):
         self.number_button9.grid(row=8, column=4, sticky=(tk.W))
         self.decimal_button.grid(row=11, column=4, sticky=(tk.W))
 
+
+        self.results_screen = results.LinksFrame(parent, self, 'test')
+
         self.reset_fields()
         # Set focus on text box
         self.UserInsert_entry.focus()
+
 
     def enable_buttons(self, enable=True):
         if enable:
@@ -220,6 +228,7 @@ class Math_Screen(tk.Frame):
         self.number_button8['state'] = set_state
         self.number_button9['state'] = set_state
         self.UserInsert_entry['state'] = set_state
+
 
     def update_time(self, start_time):
         self.time_left = start_time
@@ -285,7 +294,11 @@ class Math_Screen(tk.Frame):
                 # Check if the student's answer is correct
                 if student_answer == self.questions.answer:
                     self.Question_Count += 1
+                    self.Correct_Answers += 1
+                    # print(self.Correct_Answers)
+                    # print(self.Total_Questions)
                     self.reset_fields()
+
                 else:
                     print(f"Your answer is wrong.")
                     self.Question_Count += 1
@@ -300,16 +313,30 @@ class Math_Screen(tk.Frame):
                           font=("TkDefaultFont", 10), wraplength=101).grid(row=2, column=0, sticky=tk.W)
 
         if self.Question_Count - 1 == self.Total_Questions:
-            self.finished = True
-            self.reset_fields()
-            self.enable_buttons(False)
-            self.submit_button['state'] = 'disabled'
-            ttk.Label(self, text="You have completed all questions. Close the window to view your answers and grade.",
-                      font=("TkDefaultFont", 10), wraplength=101).grid(row=2, column=0, sticky=tk.W)
-            self.Display_Question.set('')
-            self.Question_label.set('')
+            print('u')
+
+            self.results_button = tk.Button(self, text="Show Grades",
+                                            command=lambda: results.ResultsScreen(self, 'test').mainloop())
+
+            self.results_button.grid()
+            self.reset_exercise(self.parent)
+
+
         # Set the focus back to the entry
         self.UserInsert_entry.focus()
+
+    def reset_exercise(self, parent_screen):
+        self.finished = True
+        self.reset_fields()
+        self.enable_buttons(False)
+        self.submit_button['state'] = 'disabled'
+        ttk.Label(self, text="You have completed this task.",
+                  font=("TkDefaultFont", 10), wraplength=101).grid(row=2, column=0, sticky=tk.W)
+        self.Display_Question.set('')
+        self.Question_label.set('')
+
+        # tk.Button(self, text="Start Another Task", command=lambda: parent_screen.change_screen(
+        #     parent_screen.math_problems_screen, parent_screen.problem_selection_screen)).grid()
 
     def reset_fields(self):
         self.Display_Question.set(self.questions.toggle_topics())
@@ -324,11 +351,9 @@ class Math_Screen_Settings(tk.Tk):
     def __init__(self, ID, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.geometry("600x500")
-        self.title('Math Facts Practice')
         self.resizable(width=False, height=False)
         self.ID = ID
         self.math_screen = Math_Screen(self, ID)
-
         self.math_screen.grid(sticky=(tk.E + tk.W + tk.N + tk.S))
         # self.math_screen.update_time(10)
 
@@ -359,7 +384,7 @@ class Math_Screen_Settings(tk.Tk):
 
 
 if __name__ == '__main__':
-    test_ID = '1-DIV'
+    test_ID = '1-ADD'
     app = Math_Screen_Settings(test_ID)
     # while len(app.math_screen.all_questions_list) < 3:
     app.mainloop()
