+++ src/application/views/problem_selection.py
@@ -4,8 +4,7 @@
 
 import tkinter as tk
 from tkinter import ttk
-from src.application.views import math_screen
-from src.application.views import results
+from src.application.views import math_screen, results
 
 
 #
@@ -21,10 +20,11 @@ from src.application.views import results
 # 1-DIV: Division with a single digit divisor.
 # 2-DIV: Division with a double digit divisor.
 
+
 class OptionFrame(tk.Frame):
     """A frame that contains one option for the user."""
 
-    def __init__(self, parent, name, detail, ID, screen_to_destroy, *args, **kwargs):
+    def __init__(self, parent, name, detail, ID, screen_to_destroy, master_screen, *args, **kwargs):
         """Initializes an OptionFrame widget.
         :param parent: the parent widget
         :param name: the name of the option
@@ -33,7 +33,6 @@ class OptionFrame(tk.Frame):
         super().__init__(parent, padx=20, pady=20, width=500, *args, **kwargs)
         self.name = name
         self.ID = ID
-        self.screen_to_destroy = screen_to_destroy
 
         # Create users_list LabelFrame inside the Frame
         self.label_frame = tk.LabelFrame(self, text=name, font=("TkDefaultFont", 16), padx=5, pady=5)
@@ -47,29 +46,42 @@ class OptionFrame(tk.Frame):
 
         start_frame = tk.Frame(self.label_frame, height=20)
         start_frame.grid(row=1, column=0)
-        self.start_button = ttk.Button(self.label_frame, text='Start', command=self.on_start)
+        self.start_button = ttk.Button(self.label_frame, text='Start', command=lambda: self.on_start(master_screen))
         self.start_button.place(relx=1, rely=1,
                                 anchor='se')  # Source: https://stackoverflow.com/questions/18736465/how-to-center-a-tkinter-widget
-        if self.start_is_clicked:
-            self.start_is_clicked = True
 
-    def on_start(self):
+        self.m_s = math_screen.Math_Screen(master_screen, self.ID)
+
+
+    def on_start(self, master_screen):
         """Handle the start button pressed event."""
         # Start the proper math exercise
         print(f"Starting {self.name} activity...")
         self.start_is_clicked = True
-        self.screen_to_destroy.destroy()
+        print(self.m_s)
+        master_screen.math_problems_screen[0] = self.m_s
+        master_screen.change_screen(master_screen.problem_selection_screen, master_screen.math_problems_screen)
+        print(master_screen.math_problems_screen[0])
+
+        # Results stuff. Delete if nessescary
+        #########################################################################################################
+        # results.ResultsScreen(self.m_s, 'test').mainloop()
+        if self.m_s.Question_Count - 1 == self.m_s.Total_Questions:
+            print('4')
+            tk.Button(master_screen, text="Show Grades",
+                      command=lambda: results.ResultsScreen(self, 'test').mainloop()).grid()
+        #########################################################################################################
 
 
 class SelectionView(tk.Frame):
     """The frame where the user selects which type of problems to practice."""
 
-    def __init__(self, parent, screen_to_destroy, student, *args, **kwargs):
+    def __init__(self, parent, screen_to_destroy, student, master_screen, *args, **kwargs):
         super().__init__(parent, padx=20, pady=15, *args, *kwargs)
 
         # Get grade and username
-        grade = int(student['child_grade'])
-        username = student['username']
+        self.grade = int(student['child_grade'])
+        self.username = student['username']
 
         # Create users_list toolbar with menus
         # Source: http://zetcode.com/tkinter/menustoolbars/
@@ -78,7 +90,7 @@ class SelectionView(tk.Frame):
 
         user_menu = tk.Menu(toolbar)
         user_menu.add_command(label='Log Out')
-        toolbar.add_cascade(label=username, menu=user_menu)
+        toolbar.add_cascade(label=self.username, menu=user_menu)
 
         settings_menu = tk.Menu(toolbar)
         toolbar.add_cascade(label='Settings', menu=settings_menu)
@@ -90,50 +102,46 @@ class SelectionView(tk.Frame):
         self.options = []
 
         # Use the grade to determine which tests to show
+        if self.grade == 1:
+            self.options.append(OptionFrame(self, 'Addition', 'Single digit addition.', '1-ADD', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Subtraction', 'Single digit subtraction', '1-SUB', screen_to_destroy, master_screen))
 
-        if grade == 1:
-            self.options.append(OptionFrame(self, 'Addition', 'Single digit addition.', '1-ADD', screen_to_destroy))
-            self.options.append(OptionFrame(self, 'Subtraction', 'Single digit subtraction', '1-SUB', screen_to_destroy))
-        elif grade == 2:
-            self.options.append(OptionFrame(self, 'Addition', 'Double digit addition.', '2-ADD', screen_to_destroy))
-            self.options.append(OptionFrame(self, 'Subtraction', 'Double digit subtraction.', '2-SUB', screen_to_destroy))
-        elif grade == 3:
-            self.options.append(OptionFrame(self, 'Addition', 'Double digit addition.', '2-ADD', screen_to_destroy))
-            self.options.append(OptionFrame(self, 'Subtraction', 'Double digit subtraction.', '2-SUB', screen_to_destroy))
-            self.options.append(OptionFrame(self, 'Multiplication', '0 to 12 multiplication.', '1-MUL', screen_to_destroy))
-        elif grade == 4:
-            self.options.append(OptionFrame(self, 'Addition', 'Multi-digit addition.', '3-ADD', screen_to_destroy))
+        elif self.grade == 2:
+            self.options.append(OptionFrame(self, 'Addition', 'Double digit addition.', '2-ADD', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Subtraction', 'Double digit subtraction.', '2-SUB', screen_to_destroy, master_screen))
 
-            self.options.append(
-                OptionFrame(self, 'Subtraction', 'Multi-digit subtraction.', '3-SUB', screen_to_destroy))
-            self.options.append(
-                OptionFrame(self, 'Multiplication', '0 to 12 multiplication.', '1-MUL', screen_to_destroy))
-            self.options.append(OptionFrame(self, 'Division', 'Whole number division', '1-DIV', screen_to_destroy))
+        elif self.grade == 3:
+            self.options.append(OptionFrame(self, 'Addition', 'Double digit addition.', '2-ADD', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Subtraction', 'Double digit subtraction.', '2-SUB', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Multiplication', '0 to 12 multiplication.', '1-MUL', screen_to_destroy, master_screen))
 
-        elif grade == 5:
-            self.options.append(OptionFrame(self, 'Addition', 'Multi-digit addition.', '3-ADD', screen_to_destroy))
-            self.options.append(
-                OptionFrame(self, 'Subtraction', 'Multi-digit subtraction.', '3-SUB', screen_to_destroy))
-            self.options.append(
-                OptionFrame(self, 'Multiplication', 'Double digit multiplication.', '2-MUL', screen_to_destroy))
-            self.options.append(OptionFrame(self, 'Division', 'Double digit division', '2-DIV', screen_to_destroy))
+        elif self.grade == 4:
+            self.options.append(OptionFrame(self, 'Addition', 'Multi-digit addition.', '3-ADD', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Subtraction', 'Multi-digit subtraction.', '3-SUB', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Multiplication', '0 to 12 multiplication.', '1-MUL', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Division', 'Whole number division', '1-DIV', screen_to_destroy, master_screen))
 
-        elif grade >= 6:
-            self.options.append(OptionFrame(self, 'Addition', 'Multi-digit addition.', '3-ADD', screen_to_destroy))
-            self.options.append(
-                OptionFrame(self, 'Subtraction', 'Multi-digit subtraction.', '3-SUB', screen_to_destroy))
-            self.options.append(
-                OptionFrame(self, 'Multiplication', 'Double digit multiplication.', '2-MUL', screen_to_destroy))
-            self.options.append(OptionFrame(self, 'Division', 'Double digit division', '2-DIV', screen_to_destroy))
+        elif self.grade == 5:
+            self.options.append(OptionFrame(self, 'Addition', 'Multi-digit addition.', '3-ADD', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Subtraction', 'Multi-digit subtraction.', '3-SUB', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Multiplication', 'Double digit multiplication.', '2-MUL', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Division', 'Double digit division', '2-DIV', screen_to_destroy, master_screen))
 
+        elif self.grade >= 6:
+            self.options.append(OptionFrame(self, 'Addition', 'Multi-digit addition.', '3-ADD', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Subtraction', 'Multi-digit subtraction.', '3-SUB', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Multiplication', 'Double digit multiplication.', '2-MUL', screen_to_destroy, master_screen))
+            self.options.append(OptionFrame(self, 'Division', 'Double digit division', '2-DIV', screen_to_destroy, master_screen))
             # TODO: Create a linear equations problem set for math_screen.py
             # self.options.append(OptionFrame(self, 'Algebra', 'Simple linear equations.'))
+
 
         # Place the options in the grid automatically
         max_columns = 2
 
         row = 0
         column = 0
+
         for option in self.options:
             # Add the option to the grid
             option.grid(row=row, column=column, sticky=(tk.E + tk.W))
@@ -151,15 +159,13 @@ class RootWindow(tk.Tk):
         super().__init__(*args, **kwargs)
         self.title('Math Facts Practice')
         self.resizable(width=False, height=False)
-        self.sv = SelectionView(self, self, student)
+        self.sv = SelectionView(self, self, student, self)
         self.sv.pack(expand=True, fill='both')
-
         self.columnconfigure(0, weight=1)
 
 
 def run_problem_selection(student_id, student):
     root = RootWindow(student)
-
     root.mainloop()
 
     for option in root.sv.options:
@@ -167,7 +173,7 @@ def run_problem_selection(student_id, student):
             app = math_screen.Math_Screen_Settings(option.ID)
             app.mainloop()
 
-            results_screen = results.ResultsScreen(app, student_id)
+            results_screen = results.ResultsScreen(app.math_screen, student_id)
             results_screen.mainloop()
