+++ src/application/views/welcome.py
@@ -57,7 +57,7 @@ class IconFrame(tk.Frame):
     """Contains and displays the description of the Math Facts Practice application."""
     def __init__(self, parent, **kwargs):
         super().__init__(parent, **kwargs)
-        self.image = tk.PhotoImage(file="icon.ico")
+        self.image = tk.PhotoImage(file="../tests/icon.ico")
         self.image_label = tk.Label(self, image=self.image)
         self.grid()
         self.image_label.grid()
@@ -83,14 +83,19 @@ class DescriptionFrame(tk.Frame):
 
 
 class LinksFrame(tk.Frame):
-    def __init__(self, parent, screen_to_destroy,  **kwargs):
+    def __init__(self, parent, screen_to_destroy, registration_screen, **kwargs):
         super().__init__(parent, **kwargs)
+
+        # Popup Windows
+        self.screen_to_destroy = screen_to_destroy
+        self.parent = parent
 
         # Frame with button links
         self.terms_of_use_link = ttk.Button(self, text="Terms Of Use", command=self.terms_of_use_open)
         self.terms_of_use_link.grid(row=100, column=0, sticky=tk.W)
 
-        self.registration_button = ttk.Button(self, text="Register", command=self.Registration_start)
+        self.registration_button = ttk.Button(self, text="Registration",
+            command=lambda: self.Registration_start(registration_screen))
         self.registration_button.grid(row=100, column=100, sticky=(tk.E))
         self.registration_pressed = False
 
@@ -98,9 +103,8 @@ class LinksFrame(tk.Frame):
         self.login_button.grid(row=100, column=200, sticky=(tk.E))
         self.login_pressed = False
 
-        # Popup Windows
-        self.screen_to_destroy = screen_to_destroy
-
+    def Registration_start(self, registration_screen):
+        self.screen_to_destroy.change_screen(self.parent.welcome_screen, registration_screen)
 
     def terms_of_use_open(self):
         # Terms of use window
@@ -112,21 +116,13 @@ class LinksFrame(tk.Frame):
         self.root.mainloop()
 
     def Login_start(self):
-        self.login_pressed = True
-        #login.login()
-        self.screen_to_destroy.destroy()
-        login_window = login.LoginScreen()
-        login_window.mainloop()
-
+        # self.login_pressed = True
+        # #login.login()
+        # self.screen_to_destroy.destroy()
+        # login_window = login.LoginScreen()
+        # login_window.mainloop()
+        x = 5
     # Open registration.py
-    def Registration_start(self):
-        self.registration_pressed = True
-        self.screen_to_destroy.destroy()
-        reg_screen = registration.MyApplication()
-        reg_screen.mainloop()
-
-        login_window = login.LoginScreen()
-        login_window.mainloop()
 
 
 class WelcomeView(tk.Tk):
